[WIP] Update deps libsdptransform and Catch2
- Update libsdptransform to v1.2.10
- Update Catch2 to v3.2.1

+++ scripts/get-dep.sh
@@ -16,6 +16,7 @@ function get_dep()
 	GIT_REPO="$1"
 	GIT_TAG="$2"
 	DEST="$3"
+	ADD_TO_REPO="$4"
 
 	echo ">>> [INFO] getting dep '${DEP}' ..."
 
@@ -34,9 +35,11 @@ function get_dep()
 	git checkout --quiet ${GIT_TAG}
 	set -e
 
-	echo ">>> [INFO] adding dep source code to the repository ..."
-	rm -rf .git
-	git add .
+	if [ "${ADD_TO_REPO}" == "YES" ] ; then
+		echo ">>> [INFO] adding dep source code to the repository ..."
+		rm -rf .git
+		git add .
+	fi
 
 	echo ">>> [INFO] got dep '${DEP}'"
 
@@ -46,19 +49,31 @@ function get_dep()
 function get_libsdptransform()
 {
 	GIT_REPO="https://github.com/ibc/libsdptransform.git"
-	GIT_TAG="1.2.9"
+	GIT_TAG="1.2.10"
 	DEST="deps/libsdptransform"
+	ADD_TO_REPO="YES"
 
-	get_dep "${GIT_REPO}" "${GIT_TAG}" "${DEST}"
+	get_dep "${GIT_REPO}" "${GIT_TAG}" "${DEST}" "${ADD_TO_REPO}"
 }
 
 function get_catch()
 {
 	GIT_REPO="https://github.com/philsquared/Catch.git"
-	GIT_TAG="v2.11.1"
+	GIT_TAG="v3.2.1"
 	DEST="deps/catch"
+	ADD_TO_REPO="NO"
 
-	get_dep "${GIT_REPO}" "${GIT_TAG}" "${DEST}"
+	get_dep "${GIT_REPO}" "${GIT_TAG}" "${DEST}" "${ADD_TO_REPO}"
+
+	# Catch2 v3 is no longer a single header library and we should build it as a
+	# proper (static) library. May do it in the future.
+	# Doc: https://github.com/catchorg/Catch2/blob/devel/docs/migrate-v2-to-v3.md#migrating-from-v2-to-v3
+
+	echo ">>> [INFO] copying include file to test/include/ directory ..."
+	cp ${DEST}/extras/catch_amalgamated.hpp test/include/
+
+	echo ">>> [INFO] copying source file to test/ directory ..."
+	cp ${DEST}/extras/catch_amalgamated.cpp test/src/
 }
 
 case "${DEP}" in
